# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_drafts.draftListItem:
    properties:
      created_at:
        example: "2026-06-10T12:03:00Z"
        type: string
      event_id:
        example: evt_001
        format: uuid
        type: string
      id:
        example: draft_001
        format: uuid
        type: string
      recipient:
        example: partner@company.com
        type: string
      status:
        enum:
        - pending_review
        - approved
        - sent
        example: pending_review
        type: string
      subject:
        example: Meeting Confirmation
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@mengu.ai
    name: API Support
  description: AI-assisted email automation platform. Ingests emails, analyzes them
    with LLM to extract intent and actions, then executes actions through deterministic
    handlers.
  license:
    name: Proprietary
  termsOfService: https://mengu.ai/terms
  title: Mengu AI API
  version: 1.0.0
paths:
...
  /drafts/{id}:
    get:
      description: Get a single email draft with full body content.
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              body:
                type: string
              created_at:
                type: string
              event_id:
                type: string
              id:
                type: string
              recipient:
                type: string
              status:
                type: string
              subject:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get email draft
      tags:
      - Drafts
    patch:
      consumes:
      - application/json
      description: Edit an email draft's recipient, subject, or body before approval/sending.
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      - description: Draft update fields
        in: body
        name: request
        required: true
        schema:
          properties:
            body:
              type: string
            recipient:
              type: string
            subject:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              recipient:
                type: string
              subject:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Update email draft
      tags:
      - Drafts
  /drafts/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approve a pending email draft and trigger sending.
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Approve and send email draft
      tags:
      - Drafts
...
  /events/{id}:
    get:
      description: Get a single event by ID (without analysis/logs nesting — use GET
        /events/:id for the full detail view).
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              metadata:
                type: object
              source:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get event
      tags:
      - Events
  /events/{id}/analysis:
    get:
      description: Get the AI analysis (intent, confidence, actions) for a specific
        event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              actions:
                type: array
              confidence:
                type: number
              event_id:
                type: string
              id:
                type: string
              intent:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get AI analysis for event
      tags:
      - Events
  /events/{id}/calendar-events:
    get:
      description: List calendar events created from an event (data sourced from action
        logs).
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  type: object
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - Bearer: []
      summary: Get calendar events for event
      tags:
      - Events
...
  /events/{id}/drafts:
    get:
      consumes:
      - application/json
      description: List AI-generated email drafts associated with a specific event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by status (pending_review, approved, sent)
        in: query
        name: status
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_drafts.draftListItem'
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - Bearer: []
      summary: List email drafts for event
      tags:
      - Drafts
  /events/{id}/logs:
    get:
      description: Get action execution logs (one entry per action) for a specific
        event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - Bearer: []
      summary: Get action logs for event
      tags:
      - Events
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  WebhookSecret:
    description: Webhook secret key for authenticating email webhook requests.
    in: header
    name: X-Webhook-Secret
    type: apiKey
swagger: "2.0"